requirements.txt
streamlit
pandas
requests
requests-cache
beautifulsoup4
fugashi[unidic-lite]
tqdm
python-dotenv